Utilities: Add a new class for handling of Modular integers
This class will handle cases where an integer is meant to be treated as
modular around its limit. This is useful in TCP sequence and ack numbers
because they only care about where they are relative to others in this
modular space. The main comparisons for this class check if an integer
is inside some range of modular numbers, then there are unsafe versions
of GT, LT, GE, LE for if you need to compare two of these that are
presumably close to each other.