Socket: Add support for TCP sockets
TCP sockets are now availible under SOCK_STREAM! Currently, the sockets
only have support for recieving segments (not even to a user yet).
However, you can create a socket, and then it will handle being a TCP
socket all on its own, which I am happy with for now. There are quite a
few todos of things from the spec that are not implemented yet, those
should come soon though.