Utils: Add a Timer Manager
The idea behind this class is that it can back something like a retransmission queue where there are many different lengths of timers happening at once. This functionality is different from that in the NetworkDevice where all of the timers are of one length, so a linked list works fine in managing it. This implementation is backed by an epoll instance to handle each timer expiring, and the timerfd API to do the timers. Alternatively, a priority queue would have worked, but implementing that with delete logic would have been a nightmare.